🔧 chore: metadata layout에서 page로 수정 #65
